Add ChangeVelocityCommand

+++ include/Velocity.h
@@ -9,19 +9,19 @@ class Velocity : public IProperty
 {
 public:
     Velocity() = default;
-    Velocity(int deltaX, int deltaY)
+    Velocity(double deltaX, double deltaY)
         : m_deltaX{deltaX}
         , m_deltaY{deltaY}
     {}
 
     std::string name() override { return "linear velocity"; }
 
-    int getDeltaX() { return m_deltaX; }
-    int getDeltaY() { return m_deltaY; }
+    double getDeltaX() { return m_deltaX; }
+    double getDeltaY() { return m_deltaY; }
 
 private:
-    int m_deltaX{0};
-    int m_deltaY{0};
+    double m_deltaX{0};
+    double m_deltaY{0};
 };
 
 }  // namespace otus
